Added logic for predictions
